HAL profile
