+++ .gitignore
@@ -2,4 +2,4 @@
 *~
 .vs*
 build/
-
+usrdef.h
