+++ src/fuse.c
@@ -5,9 +5,9 @@
  *        type devices that connect via USB 2.0 Full-Speed. It also has VCP-UART
  *        transfer function. It only works when installed on the AVR-DU series.
  *        Recognized by standard drivers for Windows/macos/Linux and AVRDUDE>=7.2.
- * @version 1.32.40+
- * @date 2024-07-10
- * @copyright Copyright (c) 2024 askn37 at github.com
+ * @version 1.34.49+
+ * @date 2026-08-09
+ * @copyright Copyright (c) 2026 askn37 at github.com
  * @link Product Potal : https://askn37.github.io/
  *         MIT License : https://askn37.github.io/LICENSE.html
  */
@@ -34,14 +34,14 @@
 #endif
 
 FUSES = {
-    .WDTCFG   = FUSE0_DEFAULT,
-    .BODCFG   = FUSE1_DEFAULT,
-    .OSCCFG   = FUSE2_DEFAULT,
-    .SYSCFG0  = FUSE5_DEFAULT | FUSE_EESAVE_bm | ENABLE_SYS_RESET,
-    .SYSCFG1  = FUSE6_DEFAULT,
-    .CODESIZE = FUSE7_DEFAULT,  /* 0=All application code */
-    .BOOTSIZE = FUSE8_DEFAULT,  /* 0=No bootloader */
-    .PDICFG   = FUSE10_DEFAULT  /* Never change it */
+  .WDTCFG   = FUSE0_DEFAULT,
+  .BODCFG   = FUSE1_DEFAULT,
+  .OSCCFG   = FUSE2_DEFAULT,
+  .SYSCFG0  = FUSE5_DEFAULT | FUSE_EESAVE_bm | ENABLE_SYS_RESET,
+  .SYSCFG1  = FUSE6_DEFAULT,
+  .CODESIZE = FUSE7_DEFAULT,    /* 0=All application code */
+  .BOOTSIZE = FUSE8_DEFAULT,    /* 0=No bootloader */
+  .PDICFG   = FUSE10_DEFAULT    /* Never change it */
 };
 
 // end of code
