+++ src/main.cpp
@@ -5,9 +5,9 @@
  *        type devices that connect via USB 2.0 Full-Speed. It also has VCP-UART
  *        transfer function. It only works when installed on the AVR-DU series.
  *        Recognized by standard drivers for Windows/macos/Linux and AVRDUDE>=7.2.
- * @version 1.32.40+
- * @date 2024-07-10
- * @copyright Copyright (c) 2024 askn37 at github.com
+ * @version 1.34.49+
+ * @date 2026-08-09
+ * @copyright Copyright (c) 2026 askn37 at github.com
  * @link Product Potal : https://askn37.github.io/
  *         MIT License : https://askn37.github.io/LICENSE.html
  */
@@ -17,8 +17,8 @@
 #ifndef F_CPU
   #define F_CPU 20000000L
 #endif
-#include "api/macro_api.h"  /* interrupts, initVariant */
-#include "peripheral.h"     /* import Serial (Debug) */
+#include <api/macro_api.h>  /* interrupts, initVariant */
+#include <peripheral.h>     /* import Serial (Debug) */
 #include "configuration.h"
 #include "prototype.h"
 
@@ -91,6 +91,7 @@ int main (void) {
   D1PRINTF("F_CPU = %ld\r\n", F_CPU);
   D1PRINTF("_AVR_IOXXX_H_ = " _AVR_IOXXX_H_ "\r\n");
   D1PRINTF("__AVR_ARCH__ = %d\r\n", __AVR_ARCH__);
+  D1PRINTF("HAL_PROFILE = " HAL_PROFILE "\r\n");
   DFLUSH();
 #endif
