+++ src/usb.cpp
@@ -5,9 +5,9 @@
  *        type devices that connect via USB 2.0 Full-Speed. It also has VCP-UART
  *        transfer function. It only works when installed on the AVR-DU series.
  *        Recognized by standard drivers for Windows/macos/Linux and AVRDUDE>=7.2.
- * @version 1.32.40+
- * @date 2024-07-10
- * @copyright Copyright (c) 2024 askn37 at github.com
+ * @version 1.34.49+
+ * @date 2026-08-09
+ * @copyright Copyright (c) 2026 askn37 at github.com
  * @link Product Potal : https://askn37.github.io/
  *         MIT License : https://askn37.github.io/LICENSE.html
  */
@@ -16,10 +16,10 @@
 #include <avr/eeprom.h>     /* EEMEM */
 #include <avr/pgmspace.h>   /* PROGMEM memcpy_P */
 #include <string.h>         /* memcpy */
-#include "api/macro_api.h"  /* ATOMIC_BLOCK */
-#include "api/capsule.h"    /* _CAPS macro */
-#include "api/btools.h"     /* crc32 */
-#include "peripheral.h"     /* import Serial (Debug) */
+#include <api/macro_api.h>  /* ATOMIC_BLOCK */
+#include <api/capsule.h>    /* _CAPS macro */
+#include <api/btools.h>     /* crc32 */
+#include <peripheral.h>     /* import Serial (Debug) */
 #include "configuration.h"
 #include "prototype.h"
